Protect user notes with auth middleware

+++ backend/src/controllers/notes.ts
@@ -3,11 +3,19 @@ import expressAsyncHandler from "express-async-handler";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
 import NoteModel from "../models/note";
+import { assertIsDefined } from "../utils/assertIsDefined";
 
 export const getNotes: RequestHandler = expressAsyncHandler(
   async (req, res, next) => {
+    const authenticatedUserId = req.session.userId;
+
     try {
-      const notes = await NoteModel.find().exec();
+      assertIsDefined(authenticatedUserId);
+
+      const notes = await NoteModel.find({
+        userId: authenticatedUserId,
+      }).exec();
+
       res.status(200).json(notes);
     } catch (error) {
       next(error);
@@ -17,9 +25,13 @@ export const getNotes: RequestHandler = expressAsyncHandler(
 
 export const getNote: RequestHandler = expressAsyncHandler(
   async (req, res, next) => {
+    const authenticatedUserId = req.session.userId;
+
     const noteId = req.params.noteId;
 
     try {
+      assertIsDefined(authenticatedUserId);
+
       if (!mongoose.isValidObjectId(noteId)) {
         throw createHttpError(400, "Invalid note ID!");
       }
@@ -28,6 +40,10 @@ export const getNote: RequestHandler = expressAsyncHandler(
 
       if (!note) {
         throw createHttpError(404, "Note not found!");
+      }
+
+      if (!note.userId.equals(authenticatedUserId)) {
+        throw createHttpError(401, "You are not allowed to access this note!");
       }
 
       res.status(200).json(note);
@@ -48,15 +64,20 @@ export const createNote: RequestHandler<
   CreateNoteBody,
   unknown
 > = expressAsyncHandler(async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
+
   const title = req.body.title;
   const text = req.body.text;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!title) {
       throw createHttpError(400, "Note must have a title!");
     }
 
     const newNote = await NoteModel.create({
+      userId: authenticatedUserId,
       title: title,
       text: text,
     });
@@ -81,11 +102,15 @@ export const updateNote: RequestHandler<
   UpdateNoteBody,
   unknown
 > = expressAsyncHandler(async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
+
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
 
   try {
+    assertIsDefined(authenticatedUserId);
+
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note ID!");
     }
@@ -98,6 +123,10 @@ export const updateNote: RequestHandler<
 
     if (!note) {
       throw createHttpError(404, "Note not found!");
+    }
+
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You are not allowed to access this note!");
     }
 
     note.title = newTitle;
@@ -113,9 +142,13 @@ export const updateNote: RequestHandler<
 
 export const deleteNote: RequestHandler = expressAsyncHandler(
   async (req, res, next) => {
+    const authenticatedUserId = req.session.userId;
+
     const noteId = req.params.noteId;
 
     try {
+      assertIsDefined(authenticatedUserId);
+
       if (!mongoose.isValidObjectId(noteId)) {
         throw createHttpError(400, "Invalid note ID!");
       }
@@ -124,6 +157,10 @@ export const deleteNote: RequestHandler = expressAsyncHandler(
 
       if (!note) {
         throw createHttpError(404, "Note not found!");
+      }
+
+      if (!note.userId.equals(authenticatedUserId)) {
+        throw createHttpError(401, "You are not allowed to access this note!");
       }
 
       await note.deleteOne();
